+++ token_tools/settings.py
@@ -1,0 +1,4 @@
+from django.conf import settings
+
+
+TOKEN_TIMEOUT = getattr(settings, 'TOKEN_TIMEOUT', 3600)