+++ CHANGES.md
@@ -1,3 +1,11 @@
+0.2.0
+---------------------------
+
+* Use default settings for TOKEN_TIMEOUT
+* Support Dajngo 4.1
+* Support Python 3.10
+
+
 0.1.2           (2019-12-17)
 ----------------------------